Added EndoVo as new datasource. Also, added sorting out poses and frames
Sorting of poses and frames means: There can be reasons why poses
don't exist for a frame and there can be reasons why depth maps don't
exist for a specific frame. This new code tries to ensure poses are
matched to the correct frames despite this.

Note: Somthing in the volumetric fusion does not seem right yet.
It's possible that the new sorting introduces errors, but I haven't
found them yet